requirements.txt
asyncpg
sqlalchemy
beautifulsoup4
requests
lxml
python-dotenv
aiogram